refactor: Guide 페이지 스타일 분리와 위치 변경
- Guide.styled.ts 파일 생성
- 일관성을 위해 Guide 폴더안으로 위치 이동
- styled 컴포넌트를 별도 파일로 분리하여 코드 구조 개선
