+++ src/pages/guide/Guide.tsx
@@ -1,7 +1,5 @@
 import { useState, useRef, useEffect } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
-import styled from '@emotion/styled';
-import { colors, typography } from '@/styles/global/global';
 
 import guide1 from '@/assets/signup/guide1.svg';
 import guide2 from '@/assets/signup/guide2.svg';
@@ -9,6 +7,20 @@ import guide3 from '@/assets/signup/guide3.svg';
 import guide4 from '@/assets/signup/guide4.svg';
 import guide5 from '@/assets/signup/guide5.svg';
 import guide6 from '@/assets/signup/guide6.svg';
+
+import {
+  Container,
+  Header,
+  Content,
+  Title,
+  TitleText,
+  Description,
+  MockupContainer,
+  BottomSection,
+  Indicators,
+  Indicator,
+  SkipButton,
+} from './Guide.styled';
 
 interface GuideStep {
   id: number;
@@ -197,124 +209,3 @@ const Guide = () => {
 };
 
 export default Guide;
-
-const Container = styled.div<{ isDragging?: boolean }>`
-  display: flex;
-  flex-direction: column;
-  min-width: 320px;
-  max-width: 767px;
-  min-height: 100vh;
-  margin: 0 auto;
-  align-items: center;
-  justify-content: center;
-  background: ${colors.black.main};
-  color: ${colors.white};
-  user-select: none;
-  cursor: ${({ isDragging }) => (isDragging ? 'grabbing' : 'grab')};
-`;
-
-const Header = styled.div<{ active: boolean }>`
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  z-index: 1100;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-  align-items: center;
-  max-width: 767px;
-  margin: 0 auto;
-  padding: 16px 20px;
-  background-color: ${colors.black.main};
-
-  .next-button {
-    width: 49px;
-    height: 28px;
-    padding: 4px 12px;
-    border-radius: 20px;
-    background-color: ${colors.purple.main};
-    color: ${colors.white};
-    font-size: ${typography.fontSize.sm};
-    font-weight: ${typography.fontWeight.semibold};
-    line-height: 20px;
-    text-align: center;
-    margin-left: auto;
-    cursor: pointer;
-  }
-`;
-
-const Content = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 66px 20px 0 20px;
-  gap: 20px;
-  flex: 1;
-  min-height: calc(100vh - 66px);
-`;
-
-const Title = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 135px;
-  gap: 20px;
-`;
-
-const TitleText = styled.div`
-  color: ${colors.white};
-  font-size: ${typography.fontSize['xl']};
-  font-weight: ${typography.fontWeight.bold};
-  line-height: 24px;
-`;
-
-const Description = styled.div`
-  color: ${colors.grey[100]};
-  font-size: ${typography.fontSize.base};
-  font-weight: ${typography.fontWeight.semibold};
-  line-height: 24px;
-  text-align: center;
-`;
-
-const MockupContainer = styled.div`
-  img {
-    width: 220px;
-    height: 453.052px;
-  }
-`;
-
-const BottomSection = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-direction: column;
-  gap: 20px;
-  min-width: 320px;
-  max-width: 767px;
-  padding: 0 20px;
-`;
-
-const Indicators = styled.div`
-  display: flex;
-  justify-content: center;
-  left: 50%;
-`;
-
-const Indicator = styled.div<{ active: boolean }>`
-  width: 4px;
-  height: 4px;
-  background: ${({ active }) => (active ? colors.white : colors.grey[300])};
-  border-radius: 50%;
-  margin: 0 6px;
-  cursor: pointer;
-`;
-
-const SkipButton = styled.div`
-  color: ${colors.grey[200]};
-  font-size: ${typography.fontSize.xs};
-  font-weight: ${typography.fontWeight.regular};
-  line-height: normal;
-  cursor: pointer;
-`;
